+++ src/grib_iarray.c
@@ -21,9 +21,9 @@ void grib_iarray_print(const char* title, const grib_iarray* iarray)
 {
     Assert(iarray);
     size_t i;
-    printf("%s: iarray.n=%ld  \t", title, iarray->n);
+    printf("%s: iarray.n=%lu  \t", title, (unsigned long)iarray->n);
     for (i=0; i<iarray->n; i++) {
-        printf("iarray[%ld]=%ld\t", i, iarray->v[i]);
+        printf("iarray[%lu]=%ld\t", (unsigned long)i, iarray->v[i]);
     }
     printf("\n");
 }
